adk-streaming-ws/app/requirements.txt
google-adk==1.17.0
python-dotenv>=1.0.0
